server software: handle 64-bit database IDs
The SETI@home result table is about to run out of 32-bit IDs,
so we need to move to 64-bit result IDs.
This will happen to the workunit table at some point too.

I changed the server C++ code to use the "long" type for all DB IDs
(and to use appropriate conversion codes like %lu).
"long" is 64 bit on 64-bit machines.
For uniformity I did this for all tables,
even ones (like app) that will never get big.

I chose NOT to change the DB schema for now.
The new code will work with 32-bit ID fields in the DB.
As projects approach the 32-bit limit on a table they can change
its ID field, and fields that reference this table, to BIGINT.
This is likely to happen only on the result and workunit tables.
I put functions in html/ops/db_update.php
to change the IDs of these tables.

+++ sched/update_stats.cpp
@@ -79,7 +79,7 @@ int update_users() {
         );
         retval = user.update_field(buf);
         if (retval) {
-            log_messages.printf(MSG_CRITICAL, "Can't update user %d\n", user.id);
+            log_messages.printf(MSG_CRITICAL, "Can't update user %lu\n", user.id);
             return retval;
         }
     }
@@ -112,7 +112,7 @@ int update_hosts() {
         );
         retval = host.update_field(buf);
         if (retval) {
-            log_messages.printf(MSG_CRITICAL, "Can't update host %d\n", host.id);
+            log_messages.printf(MSG_CRITICAL, "Can't update host %lu\n", host.id);
             return retval;
         }
     }
@@ -121,20 +121,20 @@ int update_hosts() {
 }
 
 int get_team_totals(TEAM& team) {
-    int nusers;
+    long nusers;
     int retval;
     DB_USER user;
     char buf[256];
 
     // count the number of users on the team
     //
-    sprintf(buf, "where teamid=%d", team.id);
+    sprintf(buf, "where teamid=%lu", team.id);
     retval = user.count(nusers, buf);
     if (retval) return retval;
 
     if (team.nusers != nusers) {
         log_messages.printf(MSG_CRITICAL,
-            "updating member count for [TEAM#%d]: database has %d users, count shows %d\n",
+            "updating member count for [TEAM#%lu]: database has %d users, count shows %ld\n",
             team.id, team.nusers, nusers
         );
     }
@@ -166,7 +166,7 @@ int update_teams() {
         retval = get_team_totals(team);
         if (retval) {
             log_messages.printf(MSG_CRITICAL,
-                "update_teams: get_team_credit([TEAM#%d]) failed: %d\n",
+                "update_teams: get_team_credit([TEAM#%lu]) failed: %d\n",
                 team.id,
                 retval
             );
@@ -184,7 +184,7 @@ int update_teams() {
         );
         retval = team.update_field(buf);
         if (retval) {
-            log_messages.printf(MSG_CRITICAL, "Can't update team %d\n", team.id);
+            log_messages.printf(MSG_CRITICAL, "Can't update team %lu\n", team.id);
             return retval;
         }
     }
